:bug: Fix migration for referential table drop (#911)

+++ prisma/migrations/20260107164151_remove_referential_table/migration.sql
@@ -1,14 +1,22 @@
 /*
+  Nota: 
+  This migration was hand-edited: it will not fail if reporting does not exist (and it usually doesn't since 
+  it's dropped by delete-views).
+
   Warnings:
 
   - You are about to drop the `Referential` table. If the table is not empty, all the data it contains will be lost.
 
 */
--- DropForeignKey
-ALTER TABLE "reporting"."Referential" DROP CONSTRAINT "Referential_dnaCode_fkey";
+DO $$
+BEGIN
+  IF EXISTS(SELECT 1 FROM information_schema.schemata WHERE schema_name = 'reporting') THEN
+    IF EXISTS(SELECT 1 FROM information_schema.tables WHERE table_schema = 'reporting' AND table_name = 'Referential') THEN
+      ALTER TABLE "reporting"."Referential" DROP CONSTRAINT IF EXISTS "Referential_dnaCode_fkey";
+      DROP TABLE "reporting"."Referential";
+    END IF;
+  END IF;
+END $$;
 
 -- DropIndex
-DROP INDEX "StructureTypologie_structureDnaCode_date_key";
-
--- DropTable
-DROP TABLE "reporting"."Referential";
+DROP INDEX IF EXISTS "StructureTypologie_structureDnaCode_date_key";
